+++ payment-email-notifications.php
@@ -31,6 +31,8 @@ require_once __DIR__ . '/functions-private.php';
 
 // Core classes.
 require_once __DIR__ . '/includes/class-plugin.php';
+require_once __DIR__ . '/includes/class-email-definitions.php';
+require_once __DIR__ . '/includes/class-settings.php';
 require_once __DIR__ . '/includes/class-status-tracker.php';
 
 /**